Health UI
Add an health interface with hearts

+++ Projet ALNS/Assets/Script/CollectionControler.cs
@@ -15,7 +15,7 @@ public class CollectionControler : MonoBehaviour
 {
 
     public Item item;
-    public float healthChange;
+    public int healthChange;
     public float moveSpeedChange;
     public float attackSpeedChange;
     public float bulletSizeChange;
